--- .eslintrc
@@ -1,22 +1,0 @@
-{
-  "extends": "airbnb",
-  "settings": {
-    "react": {
-      "version": "detect"
-    }
-  },
-  "rules": {
-    "strict": [0, "global"],
-    "func-names": 0,
-    "object-shorthand": 0,
-    "consistent-return": 0,
-    "prefer-template": 0,
-    "comma-dangle": ["error", {
-        "arrays": "always-multiline",
-        "objects": "always-multiline",
-        "imports": "always-multiline",
-        "exports": "always-multiline",
-        "functions": "never"
-    }]
-  }
-}
