+++ .prettierrc.js
@@ -1,0 +1,14 @@
+/**
+ * @see https://prettier.io/docs/en/configuration.html
+ * @type {import("prettier").Config}
+ */
+module.exports = {
+    bracketSpacing: false,
+    singleQuote: true,
+    trailingComma: 'all',
+    tabWidth: 4,
+    printWidth: 120,
+    jsxSingleQuote: true,
+    experimentalOperatorPosition: 'start',
+    objectWrap: 'collapse',
+};
