+++ jest.config.js
@@ -1,0 +1,1 @@
+module.exports = {preset: 'ts-jest', testEnvironment: 'node', testMatch: ['**/tests/**/*.ts']};
