+++ src/rules/index.ts
@@ -1,0 +1,4 @@
+import {noUnusedStyles} from './no-unused-styles';
+import {sortStyles} from './sort-styles';
+
+export const rules = {'no-unused-styles': noUnusedStyles, 'sort-styles': sortStyles};
